Add files via upload
This project retrieves data from an online JSON file and displays it in the application when a button is pressed. Upon clicking the button, a list of products is shown. Users can select a product from the list to navigate to a new page, which provides detailed information, including a description of the selected product.